Move Macros to `caliban.schema` and make it private
