+++ core/src/main/scala-3/caliban/schema/Macros.scala
@@ -1,11 +1,12 @@
-package caliban.schema.macros
+package caliban.schema
 
 import caliban.schema.Annotations.GQLExcluded
+import caliban.schema.macros.TypeInfo
 
 import scala.quoted.*
 import scala.compiletime.*
 
-object Macros {
+private object Macros {
   // this code was inspired from WIP in magnolia
   // https://github.com/propensive/magnolia/blob/b937cf2c7dabebb8236e7e948f37a354777fa9b7/src/core/macro.scala
