+++ core/src/main/scala-3/caliban/schema/SchemaDerivation.scala
@@ -6,7 +6,7 @@ import caliban.parsing.adt.Directive
 import caliban.schema.Annotations.*
 import caliban.schema.Step.{ FunctionStep, ObjectStep }
 import caliban.schema.Types.*
-import caliban.schema.macros.{ Macros, TypeInfo }
+import caliban.schema.macros.TypeInfo
 
 import scala.compiletime.*
 import scala.deriving.Mirror
